backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.9
pydantic==1.10.7
pydantic[email]
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
alembic==1.12.1 